backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
fastapi==0.129.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
smmap==5.0.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
